This check-in has more 64bit updates for built-in functions, mainly maxind and minind.
There were some "dimsizes" variables that were being malloc'ed as
ints, so I changed these to ng_size.t

_Nclmask - no change
_Nclwhere - no change
_Nclmin - no change
_Nclmax - no change
_Nclminind, _Nclmaxind - changes to return long if ind > INT32_MAX
_Nclfiledimdef - Dave fixed this one already.
_Ncldim_min - no change
_Ncldim_min_n - No change, because "n" is a dimension number, not a dimension index.
_Ncldim_max - no change
_Ncldim_max_n - No change, because "n" is a dimension number, not a dimension index.
_NclIIsInteger - no change
_NclIIsUint - no change
_NclIIsShort - no change
_NclIIsUshort - no change
_NclIIsLong - no change
_NclIIsUlong - no change
_NclIIsInt64 - no change
_NclIIsUint64 - no change
_NclIIsInt8 - no change
_NclIIsUint8 - no change
_NclIIsByte - no change
_NclIIsFloat - no change
_NclIIsDouble - no change
_NclIIsString - no change
_NclIIsChar - no change
_NclIIsNumeric - no change
_NclIIsSNumeric - no change
_NclIIsENumeric - no change
_NclIIsFile - no change
_NclIIsGraphic - no change
_NclIIsLogical - no change
_NclIFileVarTypeOf - no change
_NclITypeOf - no change

_NclIgaus - upgraded to allow nlat to be int or long, but don't allow
   nlat > INT32_MAX for now.  gaus was returning an int missing value,
   which seems wrong. I'm now having it return a double missing value
   if the input is invalid. Hopefully this won't be too backwards
   incompatible, but it seemed like a bug to me.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/branches/NCL_6_0_0@11845 10fe2890-8a40-4696-8b1d-7ccbb19c398e

+++ ni/src/ncl/AddBuiltIns.c
@@ -2123,7 +2123,7 @@ void _NclAddBuiltIns
 	nargs = 0;
 	args = NewArgs(1);
 	dimsizes[0] = 1;
-	SetArgTemplate(args,0,"integer",1,dimsizes);nargs++;
+	SetArgTemplate(args,0,"numeric",1,dimsizes);nargs++;
 	NclRegisterFunc(_NclIgaus,args,"gaus",nargs);
 	nargs = 0;
 	args = NewArgs(1);
